# src/main/resources/openapi/openapi.json
{"openapi":"3.0.2","info":{"title":"License Checker","description":"The License Checker provides information about FLOSS software licenses and helps you to identify compatible licenses.","version":"version tbd"},"paths":{"/token":{"post":{"summary":"Generate Token","operationId":"generate_token_token_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_generate_token_token_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users":{"post":{"summary":"Create User","operationId":"create_user_users_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserIn"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/me":{"get":{"summary":"Get User","operationId":"get_user_users_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/licenses/add/":{"post":{"tags":["Changing the license ontology - for admins only"],"summary":"Add License","description":"Adds a license to the ontology. Only for admins.\n\n**work in progress, not working yet**","operationId":"add_license_licenses_add__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/License"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/licenses/":{"get":{"tags":["Accessing information from the license ontology"],"summary":"Get All Licenses","description":"Returns all available software licenses as License-Objects","operationId":"get_all_licenses_licenses__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Get All Licenses Licenses  Get","type":"array","items":{"$ref":"#/components/schemas/License"}}}}}}}},"/licenses/{license_id}":{"get":{"tags":["Accessing information from the license ontology"],"summary":"Get License","description":"Returns a certain license as a License-Object","operationId":"get_license_licenses__license_id__get","parameters":[{"required":true,"schema":{"title":"License Id","type":"string"},"name":"license_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["Changing the license ontology - for admins only"],"summary":"Update License","description":"Changes the properties of a certain license. Only for admins.\n\n**work in progress, not working yet**","operationId":"update_license_licenses__license_id__put","parameters":[{"required":true,"schema":{"title":"License Id","type":"string"},"name":"license_id","in":"path"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/License"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"delete":{"tags":["Changing the license ontology - for admins only"],"summary":"Delete License","description":"Deletes a certain license. Only for admins.\n\n**work in progress, not working yet**","operationId":"delete_license_licenses__license_id__delete","parameters":[{"required":true,"schema":{"title":"License Id","type":"string"},"name":"license_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/licenses/{license_id}/conditions":{"get":{"tags":["Accessing information from the license ontology"],"summary":"Get License Conditions","description":"Returns the conditions of a certain license.","operationId":"get_license_conditions_licenses__license_id__conditions_get","parameters":[{"required":true,"schema":{"title":"License Id","type":"string"},"name":"license_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/licenses/{license_id}/limitations":{"get":{"tags":["Accessing information from the license ontology"],"summary":"Get License Limitations","description":"Returns the limitations of a certain license.","operationId":"get_license_limitations_licenses__license_id__limitations_get","parameters":[{"required":true,"schema":{"title":"License Id","type":"string"},"name":"license_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/licenses/{license_id}/permissions":{"get":{"tags":["Accessing information from the license ontology"],"summary":"Get License Permissions","description":"Returns the permissions of a certain license.","operationId":"get_license_permissions_licenses__license_id__permissions_get","parameters":[{"required":true,"schema":{"title":"License Id","type":"string"},"name":"license_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/licenses/{license_id}/license_type":{"get":{"tags":["Accessing information from the license ontology"],"summary":"Get License Type","description":"Returns the type of a certain license.","operationId":"get_license_type_licenses__license_id__license_type_get","parameters":[{"required":true,"schema":{"title":"License Id","type":"string"},"name":"license_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/licenses/{license_id}/osi_popular":{"get":{"tags":["Accessing information from the license ontology"],"summary":"Check Osi Popularity","description":"Returns TRUE if the license is among the popular licenses according to the Open Software Initiative.","operationId":"check_osi_popularity_licenses__license_id__osi_popular_get","parameters":[{"required":true,"schema":{"title":"License Id","type":"string"},"name":"license_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/popular":{"get":{"tags":["Accessing information from the license ontology"],"summary":"Return Osi Popular Licenses","description":"Returns all popular licenses according to the Open Software Initiative.","operationId":"return_osi_popular_licenses_popular_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/types/{license_type}":{"get":{"tags":["Accessing information from the license ontology"],"summary":"Return Licenses For Type","description":"Returns a list of licenses of a certain type.","operationId":"return_licenses_for_type_types__license_type__get","parameters":[{"required":true,"schema":{"title":"License Type","type":"string"},"name":"license_type","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/licenses/check/":{"post":{"tags":["Accessing information from the license ontology"],"summary":"Check License","description":"Returns which licenses are compatible with a certain license.\n\n**work in progress, not working yet**","operationId":"check_license_licenses_check__post","requestBody":{"content":{"application/json":{"schema":{"title":"Usedlicenses","type":"array","items":{"type":"string"}}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Check License Licenses Check  Post","type":"array","items":{"type":"string"}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Body_generate_token_token_post":{"title":"Body_generate_token_token_post","required":["username","password"],"type":"object","properties":{"grant_type":{"title":"Grant Type","pattern":"password","type":"string"},"username":{"title":"Username","type":"string"},"password":{"title":"Password","type":"string"},"scope":{"title":"Scope","type":"string","default":""},"client_id":{"title":"Client Id","type":"string"},"client_secret":{"title":"Client Secret","type":"string"}}},"Conditions":{"title":"Conditions","required":["id","name"],"type":"object","properties":{"id":{"title":"Id","type":"string"},"name":{"title":"Name","type":"string"}}},"HTTPValidationError":{"title":"HTTPValidationError","type":"object","properties":{"detail":{"title":"Detail","type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}}},"License":{"title":"License","required":["id","name"],"type":"object","properties":{"id":{"title":"Id","type":"string"},"name":{"title":"Name","type":"string"},"url":{"title":"Url","maxLength":65536,"minLength":1,"type":"string","format":"uri"},"type":{"$ref":"#/components/schemas/LicenseTypes"},"conditions":{"title":"Conditions","type":"array","items":{"$ref":"#/components/schemas/Conditions"},"default":[]},"permissions":{"title":"Permissions","type":"array","items":{"$ref":"#/components/schemas/Permissions"},"default":[]},"limitations":{"title":"Limitations","type":"array","items":{"$ref":"#/components/schemas/Limitations"},"default":[]},"compatibility":{"title":"Compatibility","type":"array","items":{"type":"string"},"default":[]}}},"LicenseTypes":{"title":"LicenseTypes","required":["id","name"],"type":"object","properties":{"id":{"title":"Id","type":"string"},"name":{"title":"Name","type":"string"}}},"Limitations":{"title":"Limitations","required":["id","name"],"type":"object","properties":{"id":{"title":"Id","type":"string"},"name":{"title":"Name","type":"string"}}},"Permissions":{"title":"Permissions","required":["id","name"],"type":"object","properties":{"id":{"title":"Id","type":"string"},"name":{"title":"Name","type":"string"}}},"User":{"title":"User","required":["id","username","password_hash","role"],"type":"object","properties":{"id":{"title":"Id","maximum":2147483647.0,"minimum":1.0,"type":"integer"},"username":{"title":"Username","maxLength":50,"type":"string"},"password_hash":{"title":"Password Hash","maxLength":128,"type":"string"},"role":{"title":"Role","maxLength":30,"type":"string"}},"additionalProperties":false},"UserIn":{"title":"UserIn","required":["username","password_hash","role"],"type":"object","properties":{"username":{"title":"Username","maxLength":50,"type":"string"},"password_hash":{"title":"Password Hash","maxLength":128,"type":"string"},"role":{"title":"Role","maxLength":30,"type":"string"}},"additionalProperties":false},"ValidationError":{"title":"ValidationError","required":["loc","msg","type"],"type":"object","properties":{"loc":{"title":"Location","type":"array","items":{"type":"string"}},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}}}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"token"}}}}}}